backend/requirements.txt
Django==1.7.1
Markdown==2.5.1
djangorestframework==2.4.4
pytz==2014.9
django-ipware==0.1.0
